requirements.txt
langsmith==0.1.135
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.15
langchain-ollama==0.2.0
langchain-openai==0.2.5
langchain-text-splitters==0.3.0
fastapi==0.115.2
unstructured==0.16.4
langchain-unstructured==0.1.5chromadb==0.5.13
graphrag==0.1.1
torch==2.5.1
FlagEmbedding==1.3.2
langchain-huggingface==0.1.2
weaviate-client==3.26.7
huggingface-hub==0.26.5
torchaudio==2.5.1
torchvision==0.20
transformers==4.44.2
sacremoses==0.1.1
openai==1.58.1
grandalf==0.8
langchain-chroma==0.1.4
langchain-experimental==0.3.4
duckduckgo-search==7.1.1 
docx2txt==0.8
qdrant-client==1.12.2
llama-cpp-python==0.3.5